[Refactor #131] image svg 경로 수정
