+++ store/image-store.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import { ImageState } from "../types";
 import { StaticImageData } from "next/image";
-import CamCheck from "@/assets/images/cam_check.png";
+import CamCheck from "@/assets/images/cam_check.svg";
 
 export const useCamClickImageState = create<ImageState>((set) => ({
   imageState: CamCheck,
